requirements.txt
aenum==2.2.6
alembic==1.13.1
altgraph==0.17.4
amqp==5.2.0
annotated-types==0.7.0
anyio==4.3.0
api-display-purposes==0.0.3
appdirs==1.4.4
APScheduler==3.7.0
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.8.2
blis==0.7.11
bose==2.0.22
bota==4.0.64
botasaurus==4.0.40
botasaurus_api==4.0.4
botasaurus_driver==4.0.50
botasaurus_proxy_authentication==1.0.16
botasaurus_requests==4.0.28
Brotli==1.1.0
bs4==0.0.2
catalogue==2.0.10
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
chromedriver-autoinstaller-fix==1.0.6
clarifai==10.5.0
clarifai-grpc==10.5.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
close_chrome==4.0.40
cloudpathlib==0.16.0
colorama==0.4.6
comtypes==1.4.3
confection==0.1.4
configparser==3.8.1
contextlib2==21.6.0
contourpy==1.3.0
cryptography==42.0.8
cssselect==1.2.0
cx_Freeze==7.1.0
cx_Logging==3.2.0
cycler==0.12.1
cymem==2.0.8
decorator==4.4.2
Deprecated==1.2.14
EasyProcess==1.1
emoji==1.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
facebook-sdk==3.1.0
fake-useragent==1.5.1
Faker==25.3.0
fbchat==1.9.7
Flask==2.1.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
fonttools==4.53.1
future==1.0.0
geographiclib==2.0
geopy==2.4.1
gevent==24.2.1
geventhttpclient==2.3.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huepy==1.2.1
hyperframe==6.0.1
idna==2.10
imageio==2.34.1
imageio-ffmpeg==0.4.9
importlib_metadata==7.1.0
iniconfig==2.0.0
inquirerpy==0.3.4
instabot==0.117.0
instagrapi==2.1.2
instaloader==4.11
instapy==0.6.16
itsdangerous==2.2.0
javascript_fixes==1.1.29
Jinja2==3.1.4
joblib==1.4.2
jsonschema==2.6.0
kaitaistruct==0.10
kiwisolver==1.4.7
kombu==5.3.7
langcodes==3.4.0
language_data==1.2.0
lief==0.14.1
lxml==5.2.2
lxml_html_clean==0.1.1
Mako==1.3.5
marisa-trie==1.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
MeaningCloud-python==2.0.0
mock==5.1.0
MouseInfo==0.1.3
moviepy==1.0.3
murmurhash==1.0.10
names==0.3.0
nltk==3.8.1
numpy==1.26.4
openai==0.11.0
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
paho-mqtt==1.6.1
pandas==2.2.2
pandas-stubs==2.2.2.240514
parse==1.20.1
pefile==2023.2.7
pfzy==0.3.4
pillow==10.3.0
playwright==1.44.0
pluggy==1.5.0
plyer==2.1.0
preshed==3.0.9
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==6.0.0
pyasn1==0.6.0
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydivert==2.1.0
pydub==0.25.1
pyee==8.2.2
PyGetWindow==0.0.9
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.8.0
PyMsgBox==1.0.9
pyOpenSSL==24.1.0
pyotp==2.9.0
pyparsing==3.1.2
pyperclip==1.8.2
pypiwin32==223
pyppeteer==1.0.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyquery==2.0.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
python-rapidjson==1.17
python-telegram-bot==21.2
pyttsx3==2.90
pytweening==1.2.0
pytz==2024.1
PyVirtualDisplay==3.0
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
redis==5.0.4
regex==2024.5.15
requests==2.32.2
requests-html==0.10.0
requests-toolbelt==1.0.0
responses==0.25.0
rich==13.7.1
schedule==1.2.1
schema==0.7.5
selenium==4.5.0
selenium-wire==5.1.0
semantic-version==2.10.0
setuptools==69.5.1
setuptools-rust==1.9.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.4
SQLAlchemy==1.4.41
srsly==2.4.8
tabulate==0.9.0
textblob==0.18.0.post0
thinc==8.2.3
tk==0.1.0
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
tritonclient==2.46.0
typer==0.9.4
types-pytz==2024.1.0.20240417
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==2.1
undetected-chromedriver==3.4.7
urllib3==1.26.19
vine==5.1.0
w3lib==2.2.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webdriver-manager==4.0.1
webdriverdownloader==1.1.0.4
websocket-client==1.8.0
websockets==10.4
Werkzeug==2.0.3
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
zipp==3.19.2
zope.event==5.0
zope.interface==6.4.post2
zstandard==0.22.0
apscheduler
python-telegram-bot
matplotlib